Refactoring: re-organize project intro multiple packages (folders)
