+++ .gitignore
@@ -1,1 +1,3 @@
 .idea/
+goes
+goes.exe
