+++ goes_test.go
@@ -1,4 +1,4 @@
-package goes
+package main
 
 import
 (
@@ -6,12 +6,18 @@ import
 	"github.com/satori/go.uuid"
 	"os"
 	"path"
-	"reflect"
 	"io/ioutil"
 	"bytes"
+	storage "./storage"
+	serializer "./serializer"
+	data "./data"
+	actions "./actions"
 )
 
 var tempDir string
+var handler *actions.ActionsHandler
+var _storage storage.Storage
+var _serializer serializer.Serializer
 
 type AnEvent struct {
 	A int64
@@ -26,10 +32,9 @@ type AnotherEvent struct {
 
 func setUp() {
 	tempDir := path.Join(os.TempDir(), uuid.NewV4().String())
-	storage := NewDiskStorage(tempDir)
-	SetStorage(storage)
-	serializer := NewJsonSerializer((*AnEvent)(nil), (*AnotherEvent)(nil))
-	SetSerializer(serializer)
+	_storage = storage.NewSimpleDiskStorage(tempDir)
+	_serializer = serializer.NewJsonSerializer((*AnEvent)(nil), (*AnotherEvent)(nil))
+	handler = actions.NewActionsHandler(_storage, _serializer)
 }
 
 func tearDown() {
@@ -39,8 +44,8 @@ func tearDown() {
 	}
 }
 
-func wrapEvent(aggregateId uuid.UUID, event interface{}) Event {
-	return Event{aggregateId, event}
+func wrapEvent(aggregateId uuid.UUID, event interface{}) data.Event {
+	return data.Event{AggregateId: aggregateId, Payload: event}
 }
 
 func TestSerializeEventToJson(t *testing.T) {
@@ -48,13 +53,13 @@ func TestSerializeEventToJson(t *testing.T) {
 	defer tearDown()
 
 	ev := wrapEvent(uuid.NewV4(), AnEvent{int64(1024), "Tests"})
-	err := AddEvent(ev)
+	err := handler.AddEvent(ev)
 	if err != nil {
 		t.Errorf("AddEvent failed with %q", err)
 		return
 	}
 
-	filename := (storage.(*DiskStorage)).getFilenameForEvents(ev.AggregateId.String());
+	filename := (_storage.(*storage.SimpleDiskStorage)).GetFilenameForEvents(ev.AggregateId.String());
 	if fi, _ := os.Stat(filename); fi == nil {
 		t.Errorf("AddEvent failed to create file %q", filename)
 		return
@@ -72,19 +77,19 @@ func TestSerializeEventsForSameAggregateInSameFile(t *testing.T) {
 
 	aggregateId := uuid.NewV4()
 	ev1 := wrapEvent(aggregateId, AnEvent{int64(12345), "Hello"})
-	err := AddEvent(ev1)
+	err := handler.AddEvent(ev1)
 	if err != nil {
 		t.Errorf("AddEvent failed with %q", err)
 		return
 	}
 	ev2 := wrapEvent(aggregateId, AnotherEvent{int64(23456), "Bob", 123.45})
-	err = AddEvent(ev2)
+	err = handler.AddEvent(ev2)
 	if err != nil {
 		t.Errorf("AddEvent failed with %q", err)
 		return
 	}
 
-	filename := (storage.(*DiskStorage)).getFilenameForEvents(aggregateId.String())
+	filename := (_storage.(*storage.SimpleDiskStorage)).GetFilenameForEvents(aggregateId.String())
 	content, _ := ioutil.ReadFile(filename)
 	if !bytes.Contains(content, []byte("Hello")) || !bytes.Contains(content, []byte("Bob")) {
 		t.Error("AddEvent failed. Both events are not serialized in same file.")
@@ -97,13 +102,13 @@ func TestTypeInformationIsProvided(t *testing.T) {
 	defer tearDown()
 
 	ev := wrapEvent(uuid.NewV4(), AnEvent{int64(1024), "Tests"})
-	err := AddEvent(ev)
+	err := handler.AddEvent(ev)
 	if err != nil {
 		t.Errorf("AddEvent failed with %q", err)
 		return
 	}
 
-	filename := (storage.(*DiskStorage)).getFilenameForEvents(ev.AggregateId.String());
+	filename := (_storage.(*storage.SimpleDiskStorage)).GetFilenameForEvents(ev.AggregateId.String());
 	if fi, _ := os.Stat(filename); fi == nil {
 		t.Errorf("AddEvent failed to create file %q", filename)
 		return
@@ -115,29 +120,25 @@ func TestTypeInformationIsProvided(t *testing.T) {
 	}
 }
 
-func (me *Event) Equals(other *Event) bool {
-	return me.AggregateId == other.AggregateId && reflect.DeepEqual(me.Payload, other.Payload)
-}
-
 func TestEventsCanBeRetrieved(t *testing.T) {
 	setUp()
 	defer tearDown()
 
 	aggregateId := uuid.NewV4()
 	ev1 := wrapEvent(aggregateId, AnEvent{int64(12345), "Hello"})
-	err := AddEvent(ev1)
+	err := handler.AddEvent(ev1)
 	if err != nil {
 		t.Errorf("AddEvent failed with %q", err)
 		return
 	}
 	ev2 := wrapEvent(aggregateId, AnotherEvent{int64(23456), "Bob", 123.45})
-	err = AddEvent(ev2)
+	err = handler.AddEvent(ev2)
 	if err != nil {
 		t.Errorf("AddEvent failed with %q", err)
 		return
 	}
 
-	events, err := RetrieveFor(aggregateId)
+	events, err := handler.RetrieveFor(aggregateId)
 	switch {
 	case err != nil:
 		t.Errorf("RetrieveFor(%q) failed with %q", aggregateId.String(), err)
@@ -159,11 +160,11 @@ func TestEventsCanBeReplayedInOrder(t *testing.T) {
 	testEvent1 := wrapEvent(aggregateId1, AnEvent{int64(123), "Hello 1"})
 	testEvent2 := wrapEvent(aggregateId2, AnEvent{int64(456), "Hello 2"})
 	testEvent3 := wrapEvent(aggregateId1, AnEvent{int64(789), "Hello 3"})
-	AddEvent(testEvent1)
-	AddEvent(testEvent2)
-	AddEvent(testEvent3)
+	handler.AddEvent(testEvent1)
+	handler.AddEvent(testEvent2)
+	handler.AddEvent(testEvent3)
 
-	events, err := RetrieveAll()
+	events, err := handler.RetrieveAll()
 	switch {
 	case err != nil:
 		t.Errorf("RetrieveAll failed with %q", err)
