+++ serializer/serializer.go
@@ -1,0 +1,6 @@
+package serializer
+
+type Serializer interface {
+	Serialize(interface{}) ([]byte, string, error)
+	Deserialize([]byte, string) (interface{}, error)
+}
