+++ storage/dailydisk_test.go
@@ -1,4 +1,4 @@
-package goes
+package storage
 
 import (
 	"testing"
@@ -13,7 +13,7 @@ func TestAddEvent(t *testing.T) {
 	//Arrange
 	storagePath := path.Join(os.TempDir(), uuid.NewV4().String())
 	defer os.RemoveAll(storagePath)
-	storage := NewReadableDiskStorage(storagePath)
+	storage := NewDailyDiskStorage(storagePath)
 
 	aLocation, _ := time.LoadLocation("")
 	aTime := time.Date(2016,2,11,9,53,32,1234567, aLocation)
@@ -29,7 +29,7 @@ func TestAddEvent(t *testing.T) {
 		t.Errorf("Write failed. Error: %v", err)
 	}
 
-	readableDiskStorage := storage.(*ReadableDiskStorage)
+	readableDiskStorage := storage.(*DailyDiskStorage)
 
 	globalIndexFi, _ := os.Stat(readableDiskStorage.globalIndexFilename)
 	if globalIndexFi == nil {
@@ -51,7 +51,7 @@ func TestReadStream(t *testing.T) {
 	//Arrange
 	storagePath := path.Join(os.TempDir(), uuid.NewV4().String())
 	defer os.RemoveAll(storagePath)
-	storage := NewReadableDiskStorage(storagePath)
+	storage := NewDailyDiskStorage(storagePath)
 
 	streamId := uuid.NewV4()
 	ev1 := &StoredEvent{streamId, time.Now(), "1stType", []byte("1stEvent")}
@@ -85,7 +85,7 @@ func TestReadAll(t *testing.T) {
 	//Arrange
 	storagePath := path.Join(os.TempDir(), uuid.NewV4().String())
 	defer os.RemoveAll(storagePath)
-	storage := NewReadableDiskStorage(storagePath)
+	storage := NewDailyDiskStorage(storagePath)
 
 	stream1Id := uuid.NewV4()
 	stream2Id := uuid.NewV4()
